Chore: 1차 재모듈화 준비
