+++ README.md
@@ -20,7 +20,7 @@
 ├── service
 │   ├── Authentication.py
 │   ├── InsertData.py
-│   ├── Models.py
+│   ├── Embedding_Chatbot.py
 │   ├── RedisManager.py
 │   ├── __init__.py
 │   └── __pycache__
